+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.2.1
 
+* `Pagination`
+    * `Common`:
+        * Extension `Pagination#reverse` has been added
+
 ## 0.2.0
 
 * `Repos`